Reorganize type definitions
