+++ vaadin-rich-text-editor.d.ts
@@ -1,2 +1,2 @@
 export * from './src/vaadin-rich-text-editor.js';
-export * from './@types/interfaces';
+export * from './src/interfaces';
